Fix Issue tag update to support api updates
Currently Issue api put call removes all tags if these are not passed in update params. This breaks the contract of the api because it is supposed to support partial issue updates and tags is not a required parameter. 
Put in a minor fix to update issue tags only if they are passed in params otherwise leave them untouched.